% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-package.R
\docType{package}
\name{betaSandwich}
\alias{betaSandwich}
\title{betaSandwich: Robust Confidence Intervals for Standardized Regression Coefficients}
\description{
Generates robust confidence intervals for standardized regression coefficients
using heteroscedasticity-consistent standard errors for models fitted by \code{lm()} as described in Dudgeon (2017).
}
\references{
Dudgeon, P. (2017).
Some improvements in confidence intervals for standardized regression coefficients.
\emph{Psychometrika}, \emph{82}(4), 928–951.
\doi{10.1007/s11336-017-9563-z}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{betaSandwich}
\keyword{package}
