% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods-diff-beta-sandwich.R
\name{coef.difbetasandwich}
\alias{coef.difbetasandwich}
\title{Differences of Standardized Regression Slopes}
\usage{
\method{coef}{difbetasandwich}(object, ...)
}
\arguments{
\item{object}{Object of class \code{difbetasandwich}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of differences of standardized regression slopes.
}
\description{
Differences of Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaHC(object)
out <- dif(std)
coef(out)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
