% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods-diff-beta-sandwich.R
\name{coef.diffbetasandwich}
\alias{coef.diffbetasandwich}
\title{Differences of Standardized Regression Slopes}
\usage{
\method{coef}{diffbetasandwich}(object, ...)
}
\arguments{
\item{object}{Object of class \code{diffbetasandwich}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of differences of standardized regression slopes.
}
\description{
Differences of Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaHC(object)
diff <- DiffBetaSandwich(std)
coef(diff)

}
\seealso{
Other Beta Sandwich Functions: 
\code{\link{BetaADF}()},
\code{\link{BetaHC}()},
\code{\link{BetaN}()},
\code{\link{DiffBetaSandwich}()},
\code{\link{RSqBetaSandwich}()},
\code{\link{coef.betasandwich}()},
\code{\link{coef.rsqbetasandwich}()},
\code{\link{confint.betasandwich}()},
\code{\link{confint.diffbetasandwich}()},
\code{\link{confint.rsqbetasandwich}()},
\code{\link{print.betasandwich}()},
\code{\link{print.diffbetasandwich}()},
\code{\link{print.rsqbetasandwich}()},
\code{\link{summary.betasandwich}()},
\code{\link{summary.diffbetasandwich}()},
\code{\link{summary.rsqbetasandwich}()},
\code{\link{vcov.betasandwich}()},
\code{\link{vcov.diffbetasandwich}()},
\code{\link{vcov.rsqbetasandwich}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Sandwich Functions}
\keyword{methods}
