% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods.R
\name{coef}
\alias{coef}
\title{Extract Model Coefficients}
\usage{
coef(object, ...)
}
\arguments{
\item{object}{object.}

\item{...}{additional arguments.}
}
\description{
Extract Model Coefficients
}
\keyword{methods}
