% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods.R
\name{confint.betaSandwich}
\alias{confint.betaSandwich}
\title{Robust Confidence Intervals for Standardized Regression Slopes}
\usage{
\method{confint}{betaSandwich}(object, level = 0.95, digits = 4, ...)
}
\arguments{
\item{object}{Object of class \code{betaSandwich}.}

\item{level}{the confidence level required.}

\item{digits}{Digits for rounding.}

\item{...}{additional arguments.}
}
\description{
Robust Confidence Intervals for Standardized Regression Slopes
}
\keyword{methods}
