% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods.R
\name{confint}
\alias{confint}
\title{Confidence Intervals for Model Parameters}
\usage{
confint(object, ...)
}
\arguments{
\item{object}{object.}

\item{...}{additional arguments.}
}
\description{
Confidence Intervals for Model Parameters
}
\keyword{methods}
