% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-acov-hc-dot.R
\name{.AcovHC}
\alias{.AcovHC}
\title{Asymptotic Covariance Matrix of the Standardized Parameter Vector}
\usage{
.AcovHC(jcap, gammacap, gammacap_mvn)
}
\arguments{
\item{jcap}{Numeric matrix.
Jacobian matrix of the half-vectorization
of the model-implied covariance matrix
with respect to the standardized parameter vector.}

\item{gammacap}{Numeric matrix.
Adjusted asymptotic covariance matrix.}

\item{gammacap_mvn}{Numeric matrix.
Asymptotic covariance matrix assuming multivariate normal distribution.}
}
\description{
Asymptotic Covariance Matrix of the Standardized Parameter Vector
}
\seealso{
Other BetaSandwich Functions: 
\code{\link{.CovHC}()},
\code{\link{.GammaHC}()},
\code{\link{.QMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{BetaSandwich Functions}
\keyword{acov}
\keyword{betaSandwich}
\keyword{dot}
\keyword{internal}
