% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-beta-ci.R
\name{.BetaCI}
\alias{.BetaCI}
\title{Robust Confidence Intervals for Standardized Regression Coefficients}
\usage{
.BetaCI(object, alpha = c(0.05, 0.01, 0.001))
}
\arguments{
\item{object}{Object of class \code{betaSandwich}.}
}
\value{
Returns a matrix.
}
\description{
Robust Confidence Intervals for Standardized Regression Coefficients
}
\seealso{
Other Beta Sandwich Functions: 
\code{\link{BetaHC}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Sandwich Functions}
\keyword{betaSandwich}
\keyword{ci}
\keyword{dot}
\keyword{internal}
