% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-beta-star-of-rho-dot.R
\name{.BetaStarofRho}
\alias{.BetaStarofRho}
\title{Standardized Partial Regression Slopes
(\eqn{\mathbf{P}})}
\usage{
.BetaStarofRho(rhocap, k)
}
\arguments{
\item{rhocap}{Correlation matrix.}

\item{k}{Positive integer.
Dimension of the \code{k} by \code{k} correlation matrix.}
}
\description{
Calculate standardized partial regression slopes
from the correlation matrix.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Standardized Slopes Functions}
\keyword{dot}
\keyword{internal}
\keyword{slopesstd}
\keyword{strRegression}
