% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-ci-wald-dot.R
\name{.CIWald}
\alias{.CIWald}
\title{Wald Confidence Intervals}
\usage{
.CIWald(
  est,
  se,
  theta = 0,
  alpha = c(0.05, 0.01, 0.001),
  z = FALSE,
  df,
  test = FALSE
)
}
\arguments{
\item{est}{Numeric vector.
Estimates.}

\item{se}{Numeric vector.
Standard errors of estimates.}

\item{theta}{Numeric vector.
Parameter values when the null hypothesis is true.}

\item{alpha}{Numeric vector.
Significance level/s.}

\item{z}{Logical.
If \code{z = TRUE},
use the standard normal distribution.
If \code{z = FALSE},
use the t distribution.}

\item{df}{Numeric.
Degrees of freedom if \code{z = FALSE}.}

\item{test}{Logical.
If \code{TRUE},
return only the results of hypothesis tests.
If \code{FALSE},
return both results of hypothesis tests and confidence intervals.}
}
\value{
Returns a numeric matrix with the following variables:
\describe{
\item{est}{Estimates}
\item{se}{Standard errors}
\item{t or z}{Test statistics}
\item{p}{p value}
\item{ci}{Confidence intervals}
}
Note that if \code{test = TRUE}, the \code{ci} columns are omitted.
}
\description{
Generates Wald Confidence Intervals
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{internal}
