% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-cov-hc-dot.R
\name{.CovHC}
\alias{.CovHC}
\title{Sampling Covariance Matrix of the Standardized Parameter Vector}
\usage{
.CovHC(acov, type, n, df)
}
\arguments{
\item{acov}{Numeric matrix.
Asymptotic covariance matrix of the standardized parameter vector.}

\item{type}{Character string.
Correction type.
Possible values are
\code{"hc0"},
\code{"hc1"},
\code{"hc2"},
\code{"hc3"},
\code{"hc4"},
\code{"hc4m"}, and
\code{"hc5"}.}

\item{n}{Integer.
Sample size.}

\item{k}{Integer.
\code{p} regressors plus 1.}
}
\description{
Sampling Covariance Matrix of the Standardized Parameter Vector
}
\seealso{
Other BetaSandwich Functions: 
\code{\link{.AcovHC}()},
\code{\link{.GammaHC}()},
\code{\link{.QMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{BetaSandwich Functions}
\keyword{betaSandwich}
\keyword{cov}
\keyword{dot}
\keyword{internal}
