% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-d-mat-dot.R
\name{.DMat}
\alias{.DMat}
\title{The Duplication Matrix}
\usage{
.DMat(k)
}
\arguments{
\item{k}{Positive integer.
Dimension of the \code{k} by \code{k} matrix.}
}
\value{
Returns a matrix.
}
\description{
The Duplication Matrix
}
\seealso{
Other Symmetric Functions: 
\code{\link{.PInvDmat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Symmetric Functions}
\keyword{dot}
\keyword{internal}
\keyword{linearAlgebra}
\keyword{symmetric}
