% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-d-of-mat-dot.R
\name{.DofMat}
\alias{.DofMat}
\title{Deviation from the Mean (Matrix Input)}
\usage{
.DofMat(x, center, n, k)
}
\arguments{
\item{x}{Numeric matrix.
Data matrix.}

\item{center}{Numeric vector.
Center.}

\item{n}{Positive integer.
Number of rows in the data matrix \code{x}.}

\item{k}{Positive integer.
Number of columns in the data matrix \code{x}.}
}
\value{
A matrix.
}
\description{
Deviation from the Mean (Matrix Input)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Scaling Functions}
\keyword{dot}
\keyword{internal}
\keyword{linearAlgebra}
\keyword{scaling}
