% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-gamma-hc-dot.R
\name{.GammaHC}
\alias{.GammaHC}
\title{Asymptotic Covariance Matrix of the Sample Covariance Matrix}
\usage{
.GammaHC(d, sigmacap, qcap, n)
}
\arguments{
\item{d}{Numeric matrix.
Centered data.}

\item{sigmacap}{Numeric matrix.
Covariance matrix of \eqn{Y, X_1, \dots, X_p}.}

\item{qcap}{Numeric vector
Leverage adjustment.}

\item{n}{Integer.
Sample size.}
}
\description{
Asymptotic Covariance Matrix of the Sample Covariance Matrix
}
\seealso{
Other BetaSandwich Functions: 
\code{\link{.AcovHC}()},
\code{\link{.CovHC}()},
\code{\link{.QMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{BetaSandwich Functions}
\keyword{betaSandwich}
\keyword{dot}
\keyword{gamma}
\keyword{internal}
