% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/strRegression-jacobian-vech-sigma-wrt-theta-std-dot.R
\name{.JacobianVechSigmaWRTThetaStar}
\alias{.JacobianVechSigmaWRTThetaStar}
\title{out Matrix of the Half-Vectorization
of the Model-Implied Covariance Matrix
with Respect to the Standardized Parameter Vector}
\usage{
.JacobianVechSigmaWRTThetaStar(betastar, sigmay, sigmax, rhocapx, q, p)
}
\arguments{
\item{betastar}{Numeric vector.
Standardized regression slopes.}

\item{sigmay}{Numeric.
Standard deviation of the regressand variable.}

\item{sigmax}{Numeric vector.
Standard deviations of the regressor variables.}

\item{rhocapx}{Numeric matrix.
Correlation matrix of the regressor variables.}

\item{q}{Positive integer.
Length of the parameter vector.}

\item{p}{Positive integer.
\code{p} regressors.}
}
\description{
out Matrix of the Half-Vectorization
of the Model-Implied Covariance Matrix
with Respect to the Standardized Parameter Vector
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Derivatives Functions}
\keyword{derivatives}
\keyword{dot}
\keyword{internal}
\keyword{strRegression}
