% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-moments-index-dot.R
\name{.MomentsIndex}
\alias{.MomentsIndex}
\title{Create Index for Moments Vector}
\usage{
.MomentsIndex(p)
}
\arguments{
\item{p}{Positive integer.
\code{p} regressors.}
}
\description{
Create Index for Moments Vector
}
\concept{Moments Functions}
\keyword{dot}
\keyword{internal}
\keyword{moments}
\keyword{strRegression}
