% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-pinv-of-d-mat-dot.R
\name{.PInvDmat}
\alias{.PInvDmat}
\title{The Moore-Penrose Inverse of the Duplication Matrix}
\usage{
.PInvDmat(d)
}
\arguments{
\item{d}{Numeric matrix.
Duplication matrix.}
}
\description{
The Moore-Penrose Inverse of the Duplication Matrix
}
\seealso{
Other Symmetric Functions: 
\code{\link{.DMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Symmetric Functions}
\keyword{dot}
\keyword{internal}
\keyword{linearAlgebra}
\keyword{symmetric}
