% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald-probs-of-alpha-dot.R
\name{.ProbsofAlpha}
\alias{.ProbsofAlpha}
\title{Probabilities}
\usage{
.ProbsofAlpha(alpha = c(0.05, 0.01, 0.001))
}
\arguments{
\item{alpha}{Numeric vector.
Significance level/s.}
}
\value{
Returns a vector of probabilities.
}
\description{
Generates a vector of probabilities associated with the two-tailed \code{alpha} level provided in increasing order.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{internal}
