% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-q-mat-dot.R
\name{.QMat}
\alias{.QMat}
\title{Leverage Adjustment (\eqn{\mathbf{Q}})}
\usage{
.QMat(h, k, type = "hc3", g1 = 1, g2 = 1.5, constant = 0.7)
}
\arguments{
\item{h}{Numeric vector.
Leverage values.}

\item{k}{Positive integer.
\code{p} number of regressors plus 1.}

\item{type}{Character string.
Correction type.
Possible values are
\code{"hc0"},
\code{"hc1"},
\code{"hc2"},
\code{"hc3"},
\code{"hc4"},
\code{"hc4m"}, and
\code{"hc5"}.}

\item{g1}{Numeric.
\code{g1} value for \code{type = "hc4m"} or \code{type = "hc5"}.}

\item{g2}{Numeric.
\code{g2} value for \code{type = "hc4m"}.}

\item{constant}{Numeric.
Constant for \code{type = "hc5"}}
}
\description{
Leverage Adjustment (\eqn{\mathbf{Q}})
}
\seealso{
Other BetaSandwich Functions: 
\code{\link{.AcovHC}()},
\code{\link{.CovHC}()},
\code{\link{.GammaHC}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{BetaSandwich Functions}
\keyword{betaSandwich}
\keyword{dot}
\keyword{internal}
\keyword{leverage}
