% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoMatrix-rho-of-sigma-dot.R
\name{.RhoofSigma}
\alias{.RhoofSigma}
\title{Correlation Matrix from Covariance Matrix}
\usage{
.RhoofSigma(x, q)
}
\arguments{
\item{x}{Numeric matrix.
Covariance matrix.}

\item{q}{Numeric vector.
Inverse of the standard deviation vector.}
}
\description{
Correlation Matrix from Covariance Matrix
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Correlation Functions}
\keyword{correlation}
\keyword{dot}
\keyword{internal}
\keyword{rhoMatrix}
