% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-theta-star-index-dot.R
\name{.ThetaStarIndex}
\alias{.ThetaStarIndex}
\title{Create Index for the Standardized Parameter Vector}
\usage{
.ThetaStarIndex(p)
}
\arguments{
\item{p}{Positive integer.
\code{p} regressors.}
}
\description{
Create Index for the Standardized Parameter Vector
}
\concept{Standardized Parameters Functions}
\keyword{dot}
\keyword{internal}
\keyword{parametersstd}
\keyword{strRegression}
