% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vec-dot.R
\name{.Vec}
\alias{.Vec}
\title{Vectorize a Matrix}
\usage{
.Vec(x)
}
\arguments{
\item{x}{Matrix.}
}
\value{
Returns a vector.
}
\description{
Vectorize a Matrix
}
\seealso{
Other Vectorization Functions: 
\code{\link{.VechNames}()},
\code{\link{.VechsNames}()},
\code{\link{.Vechs}()},
\code{\link{.Vech}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Vectorization Functions}
\keyword{dot}
\keyword{internal}
\keyword{linearAlgebra}
\keyword{vectorization}
