% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vech-dot.R
\name{.Vech}
\alias{.Vech}
\title{Half-Vectorize a Matrix}
\usage{
.Vech(x)
}
\arguments{
\item{x}{Matrix.}
}
\description{
Half-Vectorize a Matrix
}
\seealso{
Other Vectorization Functions: 
\code{\link{.VechNames}()},
\code{\link{.VechsNames}()},
\code{\link{.Vechs}()},
\code{\link{.Vec}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Vectorization Functions}
\keyword{dot}
\keyword{internal}
\keyword{linearAlgebra}
\keyword{vectorization}
