% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vech-names-dot.R
\name{.VechNames}
\alias{.VechNames}
\title{Vector Names for Half-Vectorization}
\usage{
.VechNames(x, sep = ".")
}
\arguments{
\item{x}{Character vector of names of length \code{k}.}

\item{sep}{Character string.
Separator for variable names.}
}
\value{
Returns a vector of character stings.
}
\description{
Vector Names for Half-Vectorization
}
\seealso{
Other Vectorization Functions: 
\code{\link{.VechsNames}()},
\code{\link{.Vechs}()},
\code{\link{.Vech}()},
\code{\link{.Vec}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Vectorization Functions}
\keyword{dot}
\keyword{internal}
\keyword{linearAlgebra}
\keyword{vectorization}
