% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vechs-dot.R
\name{.Vechs}
\alias{.Vechs}
\title{Strict Half-Vectorize a Matrix}
\usage{
.Vechs(x)
}
\arguments{
\item{x}{Matrix.}
}
\description{
Strict Half-Vectorize a Matrix
}
\seealso{
Other Vectorization Functions: 
\code{\link{.VechNames}()},
\code{\link{.VechsNames}()},
\code{\link{.Vech}()},
\code{\link{.Vec}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Vectorization Functions}
\keyword{dot}
\keyword{internal}
\keyword{linearAlgebra}
\keyword{vectorization}
