% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods-diff-beta-sandwich.R
\name{print.diffbetasandwich}
\alias{print.diffbetasandwich}
\title{Print Method for an Object of Class \code{diffbetasandwich}}
\usage{
\method{print}{diffbetasandwich}(x, alpha = 0.05, digits = 4, ...)
}
\arguments{
\item{x}{Object of class \code{diffbetasandwich}.}

\item{alpha}{Significance level.}

\item{digits}{Digits to print.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of
differences of standardized regression slopes,
standard errors,
test statistics,
degrees of freedom,
p-values,
and
confidence intervals.
}
\description{
Print Method for an Object of Class \code{diffbetasandwich}
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaHC(object)
diff <- DiffBetaSandwich(std)
print(diff)

}
\seealso{
Other Beta Sandwich Functions: 
\code{\link{BetaADF}()},
\code{\link{BetaHC}()},
\code{\link{BetaN}()},
\code{\link{DiffBetaSandwich}()},
\code{\link{RSqBetaSandwich}()},
\code{\link{coef.betasandwich}()},
\code{\link{coef.diffbetasandwich}()},
\code{\link{coef.rsqbetasandwich}()},
\code{\link{confint.betasandwich}()},
\code{\link{confint.diffbetasandwich}()},
\code{\link{confint.rsqbetasandwich}()},
\code{\link{print.betasandwich}()},
\code{\link{print.rsqbetasandwich}()},
\code{\link{summary.betasandwich}()},
\code{\link{summary.diffbetasandwich}()},
\code{\link{summary.rsqbetasandwich}()},
\code{\link{vcov.betasandwich}()},
\code{\link{vcov.diffbetasandwich}()},
\code{\link{vcov.rsqbetasandwich}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Sandwich Functions}
\keyword{methods}
