% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-stevens.R
\docType{data}
\name{stevens}
\alias{stevens}
\title{Doctoral Programs Data}
\format{
A data set from 1982 ratings of 46 doctoral programs in psychology in the USA
with the following variables:
\describe{
\item{QUALITY}{Program quality ratings.}
\item{NFACUL}{Number of faculty members in the program.}
\item{NGRADS}{Number of program graduates.}
\item{PCTSUPP}{Percentage of program graduates who received support.}
\item{PCTGRT}{Percent of faculty members holding research grants.}
\item{NARTIC}{Number of published articles attributed to program faculty member.}
\item{PCTPUB}{Percent of faculty with one or more published article.}
}
}
\usage{
stevens
}
\description{
Doctoral Programs Data
}
\references{
Stevens, J. P. (2009).
\emph{Applied multivariate statistics for the social sciences} (5th ed.).
New York: Psychology Press Taylor & Francis.
}
\keyword{data}
