% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods.R
\name{summary.betaSandwich}
\alias{summary.betaSandwich}
\title{Summary of the Results of BetaHC}
\usage{
\method{summary}{betaSandwich}(object, alpha = c(0.05, 0.01, 0.001), digits = 4, ...)
}
\arguments{
\item{object}{Object of class \code{betaSandwich}.}

\item{alpha}{Significance level.}

\item{digits}{Digits for rounding.}

\item{...}{additional arguments.}
}
\description{
Summary of the Results of BetaHC
}
\keyword{methods}
