% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods-diff-beta-sandwich.R
\name{summary.difbetasandwich}
\alias{summary.difbetasandwich}
\title{Summary Method for an Object of Class \code{difbetasandwich}}
\usage{
\method{summary}{difbetasandwich}(object, alpha = c(0.05, 0.01, 0.001), digits = 4, ...)
}
\arguments{
\item{object}{Object of class \code{difbetasandwich}.}

\item{alpha}{Significance level.}

\item{digits}{Digits to print.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of standardized regression slopes,
standard errors, test statistics, p-values, and confidence intervals.
}
\description{
Summary Method for an Object of Class \code{difbetasandwich}
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaHC(object)
out <- dif(std)
summary(out)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
