% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods-diff-beta-sandwich.R
\name{summary.diffbetasandwich}
\alias{summary.diffbetasandwich}
\title{Summary Method for an Object of Class \code{diffbetasandwich}}
\usage{
\method{summary}{diffbetasandwich}(object, alpha = NULL, digits = 4, ...)
}
\arguments{
\item{object}{Object of class \code{diffbetasandwich}.}

\item{alpha}{Numeric vector.
Significance level \eqn{\alpha}.
If \code{alpha = NULL},
use the argument \code{alpha} used in \code{object}.}

\item{digits}{Digits to print.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of
differences of standardized regression slopes,
standard errors,
test statistics,
degrees of freedom,
p-values,
and
confidence intervals.
}
\description{
Summary Method for an Object of Class \code{diffbetasandwich}
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaHC(object)
diff <- DiffBetaSandwich(std)
summary(diff)

}
\seealso{
Other Beta Sandwich Functions: 
\code{\link{BetaADF}()},
\code{\link{BetaHC}()},
\code{\link{BetaN}()},
\code{\link{DiffBetaSandwich}()},
\code{\link{RSqBetaSandwich}()},
\code{\link{coef.betasandwich}()},
\code{\link{coef.diffbetasandwich}()},
\code{\link{coef.rsqbetasandwich}()},
\code{\link{confint.betasandwich}()},
\code{\link{confint.diffbetasandwich}()},
\code{\link{confint.rsqbetasandwich}()},
\code{\link{print.betasandwich}()},
\code{\link{print.diffbetasandwich}()},
\code{\link{print.rsqbetasandwich}()},
\code{\link{summary.betasandwich}()},
\code{\link{summary.rsqbetasandwich}()},
\code{\link{vcov.betasandwich}()},
\code{\link{vcov.diffbetasandwich}()},
\code{\link{vcov.rsqbetasandwich}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Sandwich Functions}
\keyword{methods}
