% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods.R
\name{summary}
\alias{summary}
\title{Object Summaries}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{object.}

\item{...}{additional arguments.}
}
\description{
Object Summaries
}
\keyword{methods}
