% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods-beta-sandwich.R
\name{vcov.betasandwich}
\alias{vcov.betasandwich}
\title{Sampling Covariance Matrix of the Standardized Regression Slopes}
\usage{
\method{vcov}{betasandwich}(object, ...)
}
\arguments{
\item{object}{Object of class \code{betasandwich}.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of the
variance-covariance matrix
of standardized slopes.
}
\description{
Sampling Covariance Matrix of the Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaHC(object)
vcov(std)

}
\seealso{
Other Beta Sandwich Functions: 
\code{\link{BetaADF}()},
\code{\link{BetaHC}()},
\code{\link{BetaN}()},
\code{\link{DiffBetaSandwich}()},
\code{\link{RSqBetaSandwich}()},
\code{\link{coef.betasandwich}()},
\code{\link{coef.diffbetasandwich}()},
\code{\link{coef.rsqbetasandwich}()},
\code{\link{confint.betasandwich}()},
\code{\link{confint.diffbetasandwich}()},
\code{\link{confint.rsqbetasandwich}()},
\code{\link{print.betasandwich}()},
\code{\link{print.diffbetasandwich}()},
\code{\link{print.rsqbetasandwich}()},
\code{\link{summary.betasandwich}()},
\code{\link{summary.diffbetasandwich}()},
\code{\link{summary.rsqbetasandwich}()},
\code{\link{vcov.diffbetasandwich}()},
\code{\link{vcov.rsqbetasandwich}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Sandwich Functions}
\keyword{methods}
