% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods-diff-beta-sandwich.R
\name{vcov.difbetasandwich}
\alias{vcov.difbetasandwich}
\title{Sampling Covariance Matrix of
Differences of Standardized Regression Slopes}
\usage{
\method{vcov}{difbetasandwich}(object, ...)
}
\arguments{
\item{object}{Object of class \code{difbetasandwich}.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of the variance-covariance matrix
of differences of standardized regression slopes.
}
\description{
Sampling Covariance Matrix of
Differences of Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaHC(object)
out <- dif(std)
vcov(out)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
