% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods.R
\name{vcov}
\alias{vcov}
\title{Calculate Variance-Covariance Matrix for a Fitted Model Object}
\usage{
vcov(object, ...)
}
\arguments{
\item{object}{object.}

\item{...}{additional arguments.}
}
\description{
Calculate Variance-Covariance Matrix for a Fitted Model Object
}
\keyword{methods}
