% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods-r-sq-beta-sandwich.R
\name{vcov.rsqbetasandwich}
\alias{vcov.rsqbetasandwich}
\title{Sampling Covariance Matrix of
Multiple Correlation Coefficients
(R-Squared and Adjusted R-Squared)}
\usage{
\method{vcov}{rsqbetasandwich}(object, ...)
}
\arguments{
\item{object}{Object of class \code{rsqbetasandwich}.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of the
variance-covariance matrix
of multiple correlation coefficients
(R-squared and adjusted R-squared).
}
\description{
Sampling Covariance Matrix of
Multiple Correlation Coefficients
(R-Squared and Adjusted R-Squared)
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaHC(object)
rsq <- RSqBetaSandwich(std)
vcov(rsq)

}
\seealso{
Other Beta Sandwich Functions: 
\code{\link{BetaADF}()},
\code{\link{BetaHC}()},
\code{\link{BetaN}()},
\code{\link{DiffBetaSandwich}()},
\code{\link{RSqBetaSandwich}()},
\code{\link{coef.betasandwich}()},
\code{\link{coef.diffbetasandwich}()},
\code{\link{coef.rsqbetasandwich}()},
\code{\link{confint.betasandwich}()},
\code{\link{confint.diffbetasandwich}()},
\code{\link{confint.rsqbetasandwich}()},
\code{\link{print.betasandwich}()},
\code{\link{print.diffbetasandwich}()},
\code{\link{print.rsqbetasandwich}()},
\code{\link{summary.betasandwich}()},
\code{\link{summary.diffbetasandwich}()},
\code{\link{summary.rsqbetasandwich}()},
\code{\link{vcov.betasandwich}()},
\code{\link{vcov.diffbetasandwich}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Sandwich Functions}
\keyword{methods}
