% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-beta-adf.R
\name{BetaADF}
\alias{BetaADF}
\title{Estimate Standardized Regression Coefficients
and Sampling Covariance Matrix
Using the Asymptotic Distribution-Free Approach}
\usage{
BetaADF(object)
}
\arguments{
\item{object}{Object of class \code{lm}.}
}
\value{
Returns an object of class \code{betasandwich}
which is a list with the following elements:
\describe{
\item{call}{Function call.}
\item{lm}{Object of class \code{lm}.}
\item{lm_process}{Pre-processed object of class \code{lm}.}
\item{type}{Standard error type.}
\item{gamma}{Asymptotic covariance matrix
of the sample covariance matrix.}
\item{acov}{Asymptotic covariance matrix of the standardized slopes.}
\item{vcov}{Sampling covariance matrix of the standardized slopes.}
\item{est}{Vector of standardized slopes.}
}
}
\description{
Estimate Standardized Regression Coefficients
and Sampling Covariance Matrix
Using the Asymptotic Distribution-Free Approach
}
\details{
Note that while the calculation in \code{BetaADF()}
is different from \code{betaDelta::BetaDelta()} with \code{type = "adf"},
the results are numerically equivalent.
\code{BetaADF()} is appropriate when sample sizes are moderate to large
(\code{n > 250}).
\code{BetaHC()} is recommended in most situations.
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaADF(object)
# Methods ----------------------------------------------------
print(std)
summary(std)
coef(std)
vcov(std)
confint(std, level = 0.95)
## Differences of standardized regression coefficients -------
out <- dif(std)
print(out)
summary(out)
coef(out)
vcov(out)
confint(out, level = 0.95)
## Multiple Correlation --------------------------------------
out <- rsq(std)
print(out)
summary(out)
coef(out)
vcov(out)
confint(out, level = 0.95)
}
\seealso{
Other Beta Sandwich Functions: 
\code{\link{BetaHC}()},
\code{\link{BetaN}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Sandwich Functions}
\keyword{betaSandwich}
