% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-package.R
\docType{package}
\name{betaSandwich-package}
\alias{betaSandwich-package}
\alias{_PACKAGE}
\title{betaSandwich: Robust Confidence Intervals for Standardized Regression Coefficients}
\description{
Generates robust confidence intervals for standardized regression coefficients using heteroskedasticity-consistent standard errors for models fitted by lm() as described in Dudgeon (2017) \doi{10.1007/s11336-017-9563-z}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/betaSandwich}
  \item \url{https://jeksterslab.github.io/betaSandwich/}
  \item Report bugs at \url{https://github.com/jeksterslab/betaSandwich/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
