% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods.R
\name{coef.betaSandwich}
\alias{coef.betaSandwich}
\title{Standardized Regression Slopes}
\usage{
\method{coef}{betaSandwich}(object, ...)
}
\arguments{
\item{object}{Object of class \code{betaSandwich}.}

\item{...}{additional arguments.}
}
\description{
Standardized Regression Slopes
}
\keyword{methods}
