% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods.R
\name{coef.betaSandwich}
\alias{coef.betaSandwich}
\title{Standardized Regression Slopes}
\usage{
\method{coef}{betaSandwich}(object, ...)
}
\arguments{
\item{object}{Object of class \code{betaSandwich}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of standardized regression slopes.
}
\description{
Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaHC(object)
coef(std)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
