% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods-r-sq-beta-sandwich.R
\name{coef.rsqbetasandwich}
\alias{coef.rsqbetasandwich}
\title{Multiple Correlation Coefficients
(R-Squared and Adjusted R-Squared)}
\usage{
\method{coef}{rsqbetasandwich}(object, ...)
}
\arguments{
\item{object}{Object of class \code{rsqbetasandwich}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of multiple correlation coefficients
(R-squared and adjusted R-squared)
}
\description{
Multiple Correlation Coefficients
(R-Squared and Adjusted R-Squared)
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaHC(object)
rsq <- RSqBetaSandwich(std)
coef(rsq)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
