% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods-beta-sandwich.R
\name{confint.betasandwich}
\alias{confint.betasandwich}
\title{Confidence Intervals for Standardized Regression Slopes}
\usage{
\method{confint}{betasandwich}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Object of class \code{betasandwich}.}

\item{parm}{a specification of which parameters
are to be given confidence intervals,
either a vector of numbers or a vector of names.
If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of confidence intervals.
}
\description{
Confidence Intervals for Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaHC(object)
confint(std, level = 0.95)

}
\seealso{
Other Beta Sandwich Functions: 
\code{\link{BetaADF}()},
\code{\link{BetaHC}()},
\code{\link{BetaN}()},
\code{\link{DiffBetaSandwich}()},
\code{\link{RSqBetaSandwich}()},
\code{\link{coef.betasandwich}()},
\code{\link{coef.diffbetasandwich}()},
\code{\link{coef.rsqbetasandwich}()},
\code{\link{confint.diffbetasandwich}()},
\code{\link{confint.rsqbetasandwich}()},
\code{\link{print.betasandwich}()},
\code{\link{print.diffbetasandwich}()},
\code{\link{print.rsqbetasandwich}()},
\code{\link{summary.betasandwich}()},
\code{\link{summary.diffbetasandwich}()},
\code{\link{summary.rsqbetasandwich}()},
\code{\link{vcov.betasandwich}()},
\code{\link{vcov.diffbetasandwich}()},
\code{\link{vcov.rsqbetasandwich}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Sandwich Functions}
\keyword{methods}
