% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods-diff-beta-sandwich.R
\name{summary.diffbetasandwich}
\alias{summary.diffbetasandwich}
\title{Summary Method for an Object of Class \code{diffbetasandwich}}
\usage{
\method{summary}{diffbetasandwich}(object, alpha = c(0.05, 0.01, 0.001), digits = 4, ...)
}
\arguments{
\item{object}{Object of class \code{diffbetasandwich}.}

\item{alpha}{Significance level.}

\item{digits}{Digits to print.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of
differences of standardized regression slopes,
standard errors,
test statistics,
p-values,
and
confidence intervals.
}
\description{
Summary Method for an Object of Class \code{diffbetasandwich}
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaHC(object)
diff <- DiffBetaSandwich(std)
summary(diff)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
