% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods.R
\name{vcov.betaSandwich}
\alias{vcov.betaSandwich}
\title{Robust Sampling Covariance Matrix of the Standardized Regression Slopes}
\usage{
\method{vcov}{betaSandwich}(object, ...)
}
\arguments{
\item{object}{Object of class \code{betaSandwich}.}

\item{...}{additional arguments.}
}
\description{
Robust Sampling Covariance Matrix of the Standardized Regression Slopes
}
\keyword{methods}
