% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSandwich-methods.R
\name{vcov.betasandwich}
\alias{vcov.betasandwich}
\title{Robust Sampling Covariance Matrix of the Standardized Regression Slopes}
\usage{
\method{vcov}{betasandwich}(object, ...)
}
\arguments{
\item{object}{Object of class \code{betaSandwich}.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of the variance-covariance matrix
of standardized slopes.
}
\description{
Robust Sampling Covariance Matrix of the Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaHC(object)
vcov(std)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
